fix: preserve --hide-clock flag when creating new windows
The HideClock global variable was being unconditionally overwritten
by fillMissingAppearance(), which is called after the flag value
is set in run.go. This caused the clock to reappear when new windows
were spawned.

Fixed by only applying the config file value if the flag hasn't
already set it to true.